// frontend/src/api/client/models/configuration.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BotDashboard API
 * API per la gestione del Bot Dashboard con telemetria, comandi e statistiche real-time
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface Configuration {
    'k': string;
    'description'?: string | null;
    'pos'?: number;
    'value'?: string | null;
}